Migrated vulkan context to new command encoder structure
