+++ Code/Engine/RendererFoundation/CommandEncoder/Implementation/ComputeCommandEncoder.cpp
@@ -1,0 +1,49 @@
+#include <RendererFoundationPCH.h>
+
+#include <RendererFoundation/CommandEncoder/ComputeCommandEncoder.h>
+#include <RendererFoundation/Device/Device.h>
+#include <RendererFoundation/Resources/Buffer.h>
+
+void ezGALComputeCommandEncoder::Dispatch(ezUInt32 uiThreadGroupCountX, ezUInt32 uiThreadGroupCountY, ezUInt32 uiThreadGroupCountZ)
+{
+  AssertRenderingThread();
+
+  EZ_ASSERT_DEBUG(
+    uiThreadGroupCountX > 0 && uiThreadGroupCountY > 0 && uiThreadGroupCountZ > 0, "Thread group counts of zero are not meaningful. Did you mean 1?");
+
+  /// \todo Assert for compute
+
+  DispatchPlatform(uiThreadGroupCountX, uiThreadGroupCountY, uiThreadGroupCountZ);
+
+  CountDispatchCall();
+}
+
+void ezGALComputeCommandEncoder::DispatchIndirect(ezGALBufferHandle hIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes)
+{
+  AssertRenderingThread();
+  /// \todo Assert for compute
+  /// \todo Assert for indirect dispatch
+  /// \todo Assert offset < buffer size
+
+  const ezGALBuffer* pBuffer = GetDevice().GetBuffer(hIndirectArgumentBuffer);
+  EZ_ASSERT_DEV(pBuffer != nullptr, "Invalid buffer handle for indirect arguments!");
+
+  /// \todo Assert that the buffer can be used for indirect arguments (flag in desc)
+  DispatchIndirectPlatform(pBuffer, uiArgumentOffsetInBytes);
+
+  CountDispatchCall();
+}
+
+void ezGALComputeCommandEncoder::ClearStatisticsCounters()
+{
+  ezGALCommandEncoder::ClearStatisticsCounters();
+
+  m_uiDispatchCalls = 0;
+}
+
+ezGALComputeCommandEncoder::ezGALComputeCommandEncoder(ezGALDevice& device)
+  : ezGALCommandEncoder(device)
+{
+}
+
+ezGALComputeCommandEncoder::~ezGALComputeCommandEncoder() = default;
