+++ Code/Engine/RendererFoundation/CommandEncoder/RenderCommandEncoder.h
@@ -1,0 +1,106 @@
+
+#pragma once
+
+#include <Foundation/Math/Rect.h>
+#include <RendererFoundation/CommandEncoder/CommandEncoder.h>
+
+class EZ_RENDERERFOUNDATION_DLL ezGALRenderCommandEncoder : public ezGALCommandEncoder
+{
+public:
+  // Draw functions
+
+  void Draw(ezUInt32 uiVertexCount, ezUInt32 uiStartVertex);
+  void DrawIndexed(ezUInt32 uiIndexCount, ezUInt32 uiStartIndex);
+  void DrawIndexedInstanced(ezUInt32 uiIndexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartIndex);
+  void DrawIndexedInstancedIndirect(ezGALBufferHandle hIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes);
+  void DrawInstanced(ezUInt32 uiVertexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartVertex);
+  void DrawInstancedIndirect(ezGALBufferHandle hIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes);
+  void DrawAuto();
+
+  void BeginStreamOut();
+  void EndStreamOut();
+
+  // State functions
+
+  void SetIndexBuffer(ezGALBufferHandle hIndexBuffer);
+  void SetVertexBuffer(ezUInt32 uiSlot, ezGALBufferHandle hVertexBuffer);
+  void SetVertexDeclaration(ezGALVertexDeclarationHandle hVertexDeclaration);
+
+  ezGALPrimitiveTopology::Enum GetPrimitiveTopology() const { return m_Topology; }
+  void SetPrimitiveTopology(ezGALPrimitiveTopology::Enum Topology);
+
+  void SetBlendState(ezGALBlendStateHandle hBlendState, const ezColor& BlendFactor = ezColor::White, ezUInt32 uiSampleMask = 0xFFFFFFFFu);
+  void SetDepthStencilState(ezGALDepthStencilStateHandle hDepthStencilState, ezUInt8 uiStencilRefValue = 0xFFu);
+  void SetRasterizerState(ezGALRasterizerStateHandle hRasterizerState);
+
+  void SetViewport(const ezRectFloat& rect, float fMinDepth = 0.0f, float fMaxDepth = 1.0f);
+  void SetScissorRect(const ezRectU32& rect);
+
+  void SetStreamOutBuffer(ezUInt32 uiSlot, ezGALBufferHandle hBuffer, ezUInt32 uiOffset);
+
+  virtual void ClearStatisticsCounters() override;
+
+protected:
+  friend class ezGALDevice;
+
+  ezGALRenderCommandEncoder(ezGALDevice& device);
+  virtual ~ezGALRenderCommandEncoder();
+
+  // Draw functions
+
+  virtual void DrawPlatform(ezUInt32 uiVertexCount, ezUInt32 uiStartVertex) = 0;
+  virtual void DrawIndexedPlatform(ezUInt32 uiIndexCount, ezUInt32 uiStartIndex) = 0;
+  virtual void DrawIndexedInstancedPlatform(ezUInt32 uiIndexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartIndex) = 0;
+  virtual void DrawIndexedInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) = 0;
+  virtual void DrawInstancedPlatform(ezUInt32 uiVertexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartVertex) = 0;
+  virtual void DrawInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) = 0;
+  virtual void DrawAutoPlatform() = 0;
+
+  virtual void BeginStreamOutPlatform() = 0;
+  virtual void EndStreamOutPlatform() = 0;
+
+  // State functions
+
+  virtual void SetIndexBufferPlatform(const ezGALBuffer* pIndexBuffer) = 0;
+  virtual void SetVertexBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pVertexBuffer) = 0;
+  virtual void SetVertexDeclarationPlatform(const ezGALVertexDeclaration* pVertexDeclaration) = 0;
+  virtual void SetPrimitiveTopologyPlatform(ezGALPrimitiveTopology::Enum Topology) = 0;
+
+  virtual void SetBlendStatePlatform(const ezGALBlendState* pBlendState, const ezColor& BlendFactor, ezUInt32 uiSampleMask) = 0;
+  virtual void SetDepthStencilStatePlatform(const ezGALDepthStencilState* pDepthStencilState, ezUInt8 uiStencilRefValue) = 0;
+  virtual void SetRasterizerStatePlatform(const ezGALRasterizerState* pRasterizerState) = 0;
+
+  virtual void SetViewportPlatform(const ezRectFloat& rect, float fMinDepth, float fMaxDepth) = 0;
+  virtual void SetScissorRectPlatform(const ezRectU32& rect) = 0;
+
+  virtual void SetStreamOutBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pBuffer, ezUInt32 uiOffset) = 0;
+
+  virtual void InvalidateState() override;
+
+private:
+  void CountDrawCall() { m_uiDrawCalls++; }
+
+  // Used to track redundant state changes
+  ezGALBufferHandle m_hVertexBuffers[EZ_GAL_MAX_VERTEX_BUFFER_COUNT];
+  ezGALBufferHandle m_hIndexBuffer;
+
+  ezGALVertexDeclarationHandle m_hVertexDeclaration;
+  ezGALPrimitiveTopology::Enum m_Topology;
+
+  ezGALBlendStateHandle m_hBlendState;
+  ezColor m_BlendFactor;
+  ezUInt32 m_uiSampleMask;
+
+  ezGALDepthStencilStateHandle m_hDepthStencilState;
+  ezUInt8 m_uiStencilRefValue;
+
+  ezGALRasterizerStateHandle m_hRasterizerState;
+
+  ezRectU32 m_ScissorRect;
+  ezRectFloat m_ViewPortRect;
+  float m_fViewPortMinDepth;
+  float m_fViewPortMaxDepth;
+
+  // Statistic variables
+  ezUInt32 m_uiDrawCalls = 0;
+};
