+++ Code/Engine/RendererFoundation/Device/Implementation/Pass.cpp
@@ -1,16 +1,21 @@
 #include <RendererFoundationPCH.h>
 
+#include <RendererFoundation/CommandEncoder/RenderCommandEncoder.h>
 #include <RendererFoundation/Device/Pass.h>
 
-ezGALRenderCommandEncoder* ezGALPass::BeginRendering(ezGALRenderingSetup&& renderingSetup, const char* szName /*= ""*/)
+ezGALRenderCommandEncoder* ezGALPass::BeginRendering(const ezGALRenderingSetup& renderingSetup, const char* szName /*= ""*/)
 {
   EZ_ASSERT_DEV(m_CurrentCommandEncoderType == CommandEncoderType::Invalid, "Nested Command Encoder are not allowed");
   m_CurrentCommandEncoderType = CommandEncoderType::Render;
 
-  return BeginRenderingPlatform(std::move(renderingSetup), szName);
+  ezGALRenderCommandEncoder* pCommandEncoder = BeginRenderingPlatform(std::move(renderingSetup), szName);
+
+  pCommandEncoder->PushMarker(szName);
+
+  return pCommandEncoder;
 }
 
-ezGALPass::ezGALPass(ezGALDevice& device, const char* szName)
+ezGALPass::ezGALPass(ezGALDevice& device)
   : m_Device(device)
 {
 }
