+++ Code/Engine/RendererFoundation/Device/Pass.h
@@ -3,25 +3,12 @@
 
 #include <RendererFoundation/Resources/RenderTargetSetup.h>
 
-struct ezGALRenderingSetup
-{
-  ezGALRenderTargetSetup m_RenderTargetSetup;
-  ezColor m_ClearColor = ezColor(0, 0, 0, 0);
-  ezUInt32 m_uiRenderTargetClearMask = 0x0;
-  float m_fDepthClear = 1.0f;
-  ezUInt8 m_uiStencilClear = 0;
-  bool m_bClearDepth = false;
-  bool m_bClearStencil = false;
-  bool m_bDiscardColor = false;
-  bool m_bDiscardDepth = false;
-};
-
 class EZ_RENDERERFOUNDATION_DLL ezGALPass
 {
   EZ_DISALLOW_COPY_AND_ASSIGN(ezGALPass);
 
 public:
-  ezGALRenderCommandEncoder* BeginRendering(ezGALRenderingSetup&& renderingSetup, const char* szName = "");
+  ezGALRenderCommandEncoder* BeginRendering(const ezGALRenderingSetup& renderingSetup, const char* szName = "");
   void EndRendering(ezGALRenderCommandEncoder* pCommandEncoder);
 
   ezGALComputeCommandEncoder* BeginCompute(const char* szName = "");
@@ -30,13 +17,13 @@ public:
   // BeginRaytracing() could be here as well (would match Vulkan)
 
 protected:
-  virtual ezGALRenderCommandEncoder* BeginRenderingPlatform(ezGALRenderingSetup&& renderingSetup, const char* szName) = 0;
+  virtual ezGALRenderCommandEncoder* BeginRenderingPlatform(const ezGALRenderingSetup& renderingSetup, const char* szName) = 0;
   virtual void EndRenderingPlatform(ezGALRenderCommandEncoder* pCommandEncoder) = 0;
 
   virtual ezGALComputeCommandEncoder* BeginComputePlatform(const char* szName) = 0;
   virtual void EndComputePlatform(ezGALComputeCommandEncoder* pCommandEncoder) = 0;
 
-  ezGALPass(ezGALDevice& device, const char* szName);
+  ezGALPass(ezGALDevice& device);
   virtual ~ezGALPass();
 
   ezGALDevice& m_Device;
