+++ Code/Engine/RendererFoundation/Resources/RenderTargetSetup.h
@@ -32,4 +32,17 @@ protected:
   ezUInt8 m_uiMaxRTIndex;
 };
 
+struct ezGALRenderingSetup
+{
+  ezGALRenderTargetSetup m_RenderTargetSetup;
+  ezColor m_ClearColor = ezColor(0, 0, 0, 0);
+  ezUInt32 m_uiRenderTargetClearMask = 0x0;
+  float m_fDepthClear = 1.0f;
+  ezUInt8 m_uiStencilClear = 0;
+  bool m_bClearDepth = false;
+  bool m_bClearStencil = false;
+  bool m_bDiscardColor = false;
+  bool m_bDiscardDepth = false;
+};
+
 #include <RendererFoundation/Resources/Implementation/RenderTargetSetup_inl.h>
