+++ Code/Engine/RendererVulkan/CommandEncoder/CommandEncoderVulkan.h
@@ -2,7 +2,7 @@
 #pragma once
 
 #include <Foundation/Types/Bitflags.h>
-#include <RendererFoundation/Context/Context.h>
+#include <RendererFoundation/CommandEncoder/CommandEncoder.h>
 #include <RendererFoundation/Device/Device.h>
 #include <RendererVulkan/RendererVulkanDLL.h>
 
@@ -12,103 +12,40 @@ class ezGALBlendStateVulkan;
 class ezGALBufferVulkan;
 class ezGALDepthStencilStateVulkan;
 class ezGALRasterizerStateVulkan;
-class ezGALRenderTargetViewVulkan;
 class ezGALResourceViewVulkan;
 class ezGALSamplerStateVulkan;
 class ezGALShaderVulkan;
 class ezGALUnorderedAccessViewVulkan;
 
-/// \brief The Vulkan implementation of the graphics context.
-class EZ_RENDERERVULKAN_DLL ezGALContextVulkan : public ezGALContext
+/// \brief The Vulkan implementation of the command encoder.
+template <typename Base>
+class EZ_RENDERERVULKAN_DLL ezGALCommandEncoderVulkan : public Base
 {
 public:
 protected:
   friend class ezGALDeviceVulkan;
   friend class ezMemoryUtils;
 
-  ezGALContextVulkan(ezGALDevice* pDevice);
-
-  ~ezGALContextVulkan();
-
-  // Draw functions
-
-  virtual void ClearPlatform(const ezColor& ClearColor, ezUInt32 uiRenderTargetClearMask, bool bClearDepth, bool bClearStencil, float fDepthClear, ezUInt8 uiStencilClear) override;
-
-  virtual void ClearUnorderedAccessViewPlatform(const ezGALUnorderedAccessView* pUnorderedAccessView, ezVec4 clearValues) override;
-
-  virtual void ClearUnorderedAccessViewPlatform(const ezGALUnorderedAccessView* pUnorderedAccessView, ezVec4U32 clearValues) override;
-
-  virtual void DrawPlatform(ezUInt32 uiVertexCount, ezUInt32 uiStartVertex) override;
-
-  virtual void DrawIndexedPlatform(ezUInt32 uiIndexCount, ezUInt32 uiStartIndex) override;
-
-  virtual void DrawIndexedInstancedPlatform(ezUInt32 uiIndexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartIndex) override;
-
-  virtual void DrawIndexedInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
-
-  virtual void DrawInstancedPlatform(ezUInt32 uiVertexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartVertex) override;
-
-  virtual void DrawInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
-
-  virtual void DrawAutoPlatform() override;
-
-  virtual void BeginStreamOutPlatform() override;
-
-  virtual void EndStreamOutPlatform() override;
-
-  // Dispatch
-
-  virtual void DispatchPlatform(ezUInt32 uiThreadGroupCountX, ezUInt32 uiThreadGroupCountY, ezUInt32 uiThreadGroupCountZ) override;
-
-  virtual void DispatchIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
-
+  ezGALCommandEncoderVulkan(ezGALDevice& device);
+  ~ezGALCommandEncoderVulkan();
 
   // State setting functions
 
   virtual void SetShaderPlatform(const ezGALShader* pShader) override;
 
-  virtual void SetIndexBufferPlatform(const ezGALBuffer* pIndexBuffer) override;
-
-  virtual void SetVertexBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pVertexBuffer) override;
-
-  virtual void SetVertexDeclarationPlatform(const ezGALVertexDeclaration* pVertexDeclaration) override;
-
-  virtual void SetPrimitiveTopologyPlatform(ezGALPrimitiveTopology::Enum Topology) override;
-
   virtual void SetConstantBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pBuffer) override;
-
   virtual void SetSamplerStatePlatform(ezGALShaderStage::Enum Stage, ezUInt32 uiSlot, const ezGALSamplerState* pSamplerState) override;
-
   virtual void SetResourceViewPlatform(ezGALShaderStage::Enum Stage, ezUInt32 uiSlot, const ezGALResourceView* pResourceView) override;
-
-  virtual void SetRenderTargetSetupPlatform(ezArrayPtr<const ezGALRenderTargetView*> pRenderTargetViews, const ezGALRenderTargetView* pDepthStencilView) override;
-
   virtual void SetUnorderedAccessViewPlatform(ezUInt32 uiSlot, const ezGALUnorderedAccessView* pUnorderedAccessView) override;
-
-  virtual void SetBlendStatePlatform(const ezGALBlendState* pBlendState, const ezColor& BlendFactor, ezUInt32 uiSampleMask) override;
-
-  virtual void SetDepthStencilStatePlatform(const ezGALDepthStencilState* pDepthStencilState, ezUInt8 uiStencilRefValue) override;
-
-  virtual void SetRasterizerStatePlatform(const ezGALRasterizerState* pRasterizerState) override;
-
-  virtual void SetViewportPlatform(const ezRectFloat& rect, float fMinDepth, float fMaxDepth) override;
-
-  virtual void SetScissorRectPlatform(const ezRectU32& rect) override;
-
-  virtual void SetStreamOutBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pBuffer, ezUInt32 uiOffset) override;
 
   // Fence & Query functions
 
   virtual void InsertFencePlatform(const ezGALFence* pFence) override;
-
   virtual bool IsFenceReachedPlatform(const ezGALFence* pFence) override;
-
   virtual void WaitForFencePlatform(const ezGALFence* pFence) override;
 
   virtual void BeginQueryPlatform(const ezGALQuery* pQuery) override;
-
   virtual void EndQueryPlatform(const ezGALQuery* pQuery) override;
-
   virtual ezResult GetQueryResultPlatform(const ezGALQuery* pQuery, ezUInt64& uiQueryResult) override;
 
   // Timestamp functions
@@ -117,14 +54,15 @@ protected:
 
   // Resource update functions
 
-  virtual void CopyBufferPlatform(const ezGALBuffer* pDestination, const ezGALBuffer* pSource) override;
+  virtual void ClearUnorderedAccessViewPlatform(const ezGALUnorderedAccessView* pUnorderedAccessView, ezVec4 clearValues) override;
+  virtual void ClearUnorderedAccessViewPlatform(const ezGALUnorderedAccessView* pUnorderedAccessView, ezVec4U32 clearValues) override;
 
+  virtual void CopyBufferPlatform(const ezGALBuffer* pDestination, const ezGALBuffer* pSource) override;
   virtual void CopyBufferRegionPlatform(const ezGALBuffer* pDestination, ezUInt32 uiDestOffset, const ezGALBuffer* pSource, ezUInt32 uiSourceOffset, ezUInt32 uiByteCount) override;
 
   virtual void UpdateBufferPlatform(const ezGALBuffer* pDestination, ezUInt32 uiDestOffset, ezArrayPtr<const ezUInt8> pSourceData, ezGALUpdateMode::Enum updateMode) override;
 
   virtual void CopyTexturePlatform(const ezGALTexture* pDestination, const ezGALTexture* pSource) override;
-
   virtual void CopyTextureRegionPlatform(const ezGALTexture* pDestination, const ezGALTextureSubresource& DestinationSubResource, const ezVec3U32& DestinationPoint, const ezGALTexture* pSource, const ezGALTextureSubresource& SourceSubResource, const ezBoundingBoxu32& Box) override;
 
   virtual void UpdateTexturePlatform(const ezGALTexture* pDestination, const ezGALTextureSubresource& DestinationSubResource, const ezBoundingBoxu32& DestinationBox, const ezGALSystemMemoryDescription& pSourceData) override;
@@ -144,21 +82,15 @@ protected:
   // Debug helper functions
 
   virtual void PushMarkerPlatform(const char* szMarker) override;
-
   virtual void PopMarkerPlatform() override;
-
   virtual void InsertEventMarkerPlatform(const char* szMarker) override;
 
   void FlushDeferredStateChanges();
 
-  vk::Device m_device;
-  vk::CommandPool m_commandPool;
+  vk::Device m_vkDevice;
 
-  static constexpr ezUInt32 NUM_CMD_BUFFERS = 4;
-  vk::CommandBuffer m_commandBuffers[NUM_CMD_BUFFERS];
-  vk::Fence m_commandBufferFences[NUM_CMD_BUFFERS];
+  vk::CommandBuffer* m_pCommandBuffer = nullptr;
 
-  ezUInt32 m_uiCurrentCmdBufferIndex = 0;
   const ezGALShaderVulkan* m_pCurrentShader;
   const ezGALBlendStateVulkan* m_pCurrentBlendState;
   const ezGALDepthStencilStateVulkan* m_pCurrentDepthStencilState;
@@ -194,4 +126,4 @@ protected:
   ezGAL::ModifiedRange m_BoundSamplerStatesRange[ezGALShaderStage::ENUM_COUNT];
 };
 
-#include <RendererVulkan/Context/Implementation/ContextVulkan_inl.h>
+#include <RendererVulkan/CommandEncoder/Implementation/CommandEncoderVulkan_inl.h>
