+++ Code/Engine/RendererVulkan/CommandEncoder/RenderCommandEncoderVulkan.h
@@ -1,0 +1,53 @@
+
+#pragma once
+
+#include <RendererFoundation/CommandEncoder/RenderCommandEncoder.h>
+#include <RendererFoundation/Resources/RenderTargetSetup.h>
+#include <RendererVulkan/CommandEncoder/CommandEncoderVulkan.h>
+
+class EZ_RENDERERVULKAN_DLL ezGALRenderCommandEncoderVulkan : public ezGALCommandEncoderVulkan<ezGALRenderCommandEncoder>
+{
+protected:
+  friend class ezGALDeviceVulkan;
+  friend class ezGALPassVulkan;
+  friend class ezMemoryUtils;
+
+  using SUPER = ezGALCommandEncoderVulkan<ezGALRenderCommandEncoder>;
+
+  ezGALRenderCommandEncoderVulkan(ezGALDevice& device);
+  virtual ~ezGALRenderCommandEncoderVulkan();
+
+  void BeginEncode(vk::CommandBuffer& commandBuffer, const ezGALRenderingSetup& renderingSetup);
+  void EndEncode();
+
+  // Draw functions
+
+  virtual void DrawPlatform(ezUInt32 uiVertexCount, ezUInt32 uiStartVertex) override;
+  virtual void DrawIndexedPlatform(ezUInt32 uiIndexCount, ezUInt32 uiStartIndex) override;
+  virtual void DrawIndexedInstancedPlatform(ezUInt32 uiIndexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartIndex) override;
+  virtual void DrawIndexedInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
+  virtual void DrawInstancedPlatform(ezUInt32 uiVertexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartVertex) override;
+  virtual void DrawInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
+  virtual void DrawAutoPlatform() override;
+
+  virtual void BeginStreamOutPlatform() override;
+  virtual void EndStreamOutPlatform() override;
+
+  // State functions
+
+  virtual void SetIndexBufferPlatform(const ezGALBuffer* pIndexBuffer) override;
+  virtual void SetVertexBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pVertexBuffer) override;
+  virtual void SetVertexDeclarationPlatform(const ezGALVertexDeclaration* pVertexDeclaration) override;
+  virtual void SetPrimitiveTopologyPlatform(ezGALPrimitiveTopology::Enum Topology) override;
+
+  virtual void SetBlendStatePlatform(const ezGALBlendState* pBlendState, const ezColor& BlendFactor, ezUInt32 uiSampleMask) override;
+  virtual void SetDepthStencilStatePlatform(const ezGALDepthStencilState* pDepthStencilState, ezUInt8 uiStencilRefValue) override;
+  virtual void SetRasterizerStatePlatform(const ezGALRasterizerState* pRasterizerState) override;
+
+  virtual void SetViewportPlatform(const ezRectFloat& rect, float fMinDepth, float fMaxDepth) override;
+  virtual void SetScissorRectPlatform(const ezRectU32& rect) override;
+
+  virtual void SetStreamOutBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pBuffer, ezUInt32 uiOffset) override;
+
+private:
+};
