+++ Code/Engine/RendererVulkan/Device/DeviceVulkan.h
@@ -2,6 +2,7 @@
 #pragma once
 
 #include <Foundation/Types/Bitflags.h>
+#include <Foundation/Types/UniquePtr.h>
 #include <RendererFoundation/Device/Device.h>
 #include <RendererVulkan/RendererVulkanDLL.h>
 
@@ -12,6 +13,7 @@ using ezGALFormatLookupTableVulkan = ezGALFormatLookupTable<ezGALFormatLookupEnt
 
 class ezGALBufferVulkan;
 class ezGALTextureVulkan;
+class ezGALPassVulkan;
 
 /// \brief The Vulkan device implementation of the graphics abstraction layer.
 class EZ_RENDERERVULKAN_DLL ezGALDeviceVulkan : public ezGALDevice
@@ -28,6 +30,8 @@ public:
   vk::Instance GetVulkanInstance() const;
   vk::Device GetVulkanDevice() const;
 
+  vk::CommandBuffer& GetPrimaryCommandBuffer();
+
   ezArrayPtr<const ezUInt32> GetQueueFamilyIndices() const;
   vk::Queue GetQueue();
 
@@ -36,6 +40,9 @@ public:
   ezInt32 GetMemoryIndex(vk::MemoryPropertyFlags properties, const vk::MemoryRequirements& requirements) const;
 
   void ReportLiveGpuObjects();
+
+  ezGALBufferVulkan* FindTempBuffer(ezUInt32 uiSize) { return nullptr; }                                                                           // TODO impl
+  ezGALTextureVulkan* FindTempTexture(ezUInt32 uiWidth, ezUInt32 uiHeight, ezUInt32 uiDepth, ezGALResourceFormat::Enum format) { return nullptr; } // TODO impl
 
   // These functions need to be implemented by a render API abstraction
 protected:
@@ -48,77 +55,66 @@ protected:
   //ezResult InitPlatform(DWORD flags, IDXGIAdapter* pUsedAdapter);
 
   virtual ezResult InitPlatform() override;
-
   virtual ezResult ShutdownPlatform() override;
+
+  // Pass functions
+
+  virtual ezGALPass* BeginPassPlatform(const char* szName) override;
+  virtual void EndPassPlatform(ezGALPass* pPass) override;
 
 
   // State creation functions
 
   virtual ezGALBlendState* CreateBlendStatePlatform(const ezGALBlendStateCreationDescription& Description) override;
-
   virtual void DestroyBlendStatePlatform(ezGALBlendState* pBlendState) override;
 
   virtual ezGALDepthStencilState* CreateDepthStencilStatePlatform(const ezGALDepthStencilStateCreationDescription& Description) override;
-
   virtual void DestroyDepthStencilStatePlatform(ezGALDepthStencilState* pDepthStencilState) override;
 
   virtual ezGALRasterizerState* CreateRasterizerStatePlatform(const ezGALRasterizerStateCreationDescription& Description) override;
-
   virtual void DestroyRasterizerStatePlatform(ezGALRasterizerState* pRasterizerState) override;
 
   virtual ezGALSamplerState* CreateSamplerStatePlatform(const ezGALSamplerStateCreationDescription& Description) override;
-
   virtual void DestroySamplerStatePlatform(ezGALSamplerState* pSamplerState) override;
 
 
   // Resource creation functions
 
   virtual ezGALShader* CreateShaderPlatform(const ezGALShaderCreationDescription& Description) override;
-
   virtual void DestroyShaderPlatform(ezGALShader* pShader) override;
 
   virtual ezGALBuffer* CreateBufferPlatform(const ezGALBufferCreationDescription& Description, ezArrayPtr<const ezUInt8> pInitialData) override;
-
   virtual void DestroyBufferPlatform(ezGALBuffer* pBuffer) override;
 
   virtual ezGALTexture* CreateTexturePlatform(const ezGALTextureCreationDescription& Description, ezArrayPtr<ezGALSystemMemoryDescription> pInitialData) override;
-
   virtual void DestroyTexturePlatform(ezGALTexture* pTexture) override;
 
   virtual ezGALResourceView* CreateResourceViewPlatform(ezGALResourceBase* pResource, const ezGALResourceViewCreationDescription& Description) override;
-
   virtual void DestroyResourceViewPlatform(ezGALResourceView* pResourceView) override;
 
   virtual ezGALRenderTargetView* CreateRenderTargetViewPlatform(ezGALTexture* pTexture, const ezGALRenderTargetViewCreationDescription& Description) override;
-
   virtual void DestroyRenderTargetViewPlatform(ezGALRenderTargetView* pRenderTargetView) override;
 
   ezGALUnorderedAccessView* CreateUnorderedAccessViewPlatform(ezGALResourceBase* pResource, const ezGALUnorderedAccessViewCreationDescription& Description) override;
-
   virtual void DestroyUnorderedAccessViewPlatform(ezGALUnorderedAccessView* pResource) override;
 
   // Other rendering creation functions
 
   virtual ezGALSwapChain* CreateSwapChainPlatform(const ezGALSwapChainCreationDescription& Description) override;
-
   virtual void DestroySwapChainPlatform(ezGALSwapChain* pSwapChain) override;
 
   virtual ezGALFence* CreateFencePlatform() override;
-
   virtual void DestroyFencePlatform(ezGALFence* pFence) override;
 
   virtual ezGALQuery* CreateQueryPlatform(const ezGALQueryCreationDescription& Description) override;
-
   virtual void DestroyQueryPlatform(ezGALQuery* pQuery) override;
 
   virtual ezGALVertexDeclaration* CreateVertexDeclarationPlatform(const ezGALVertexDeclarationCreationDescription& Description) override;
-
   virtual void DestroyVertexDeclarationPlatform(ezGALVertexDeclaration* pVertexDeclaration) override;
 
   // Timestamp functions
 
   virtual ezGALTimestampHandle GetTimestampPlatform() override;
-
   virtual ezResult GetTimestampResultPlatform(ezGALTimestampHandle hTimestamp, ezTime& result) override;
 
   // Swap chain functions
@@ -128,7 +124,6 @@ protected:
   // Misc functions
 
   virtual void BeginFramePlatform() override;
-
   virtual void EndFramePlatform() override;
 
   virtual void SetPrimarySwapChainPlatform(ezGALSwapChain* pSwapChain) override;
@@ -138,8 +133,6 @@ protected:
   /// \endcond
 
 private:
-  friend class ezGALContextVulkan;
-
   struct TempResourceType
   {
     enum Enum
@@ -159,9 +152,6 @@ private:
     ezUInt64 m_uiFrame = -1;
   };
 
-
-  ezGALBufferVulkan* FindTempBuffer(ezUInt32 uiSize) { return nullptr; }                                                                           // TODO impl
-  ezGALTextureVulkan* FindTempTexture(ezUInt32 uiWidth, ezUInt32 uiHeight, ezUInt32 uiDepth, ezGALResourceFormat::Enum format) { return nullptr; } // TODO impl
   void FreeTempResources(ezUInt64 uiFrame);
 
   //ID3D11Query* GetTimestamp(ezGALTimestampHandle hTimestamp);
@@ -177,6 +167,15 @@ private:
   ezGALFormatLookupTableVulkan m_FormatLookupTable;
 
   vk::PhysicalDeviceMemoryProperties m_memoryProperties;
+
+  vk::CommandPool m_commandPool;
+
+  static constexpr ezUInt32 NUM_CMD_BUFFERS = 4;
+  vk::CommandBuffer m_commandBuffers[NUM_CMD_BUFFERS];
+  vk::Fence m_commandBufferFences[NUM_CMD_BUFFERS];
+  ezUInt32 m_uiCurrentCmdBufferIndex = 0;
+
+  ezUniquePtr<ezGALPassVulkan> m_pDefaultPass;
 
   PerFrameData m_PerFrameData[4];
   ezUInt8 m_uiCurrentPerFrameData = 0;
